requirements.txt

bcrypt==4.1.2; python_version >= '3.7'
cffi==1.16.0; python_version >= '3.8'
cryptography==41.0.7
dnspython==2.4.2; python_version >= '3.8' and python_version < '4.0'
ecdsa==0.18.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
email-validator==2.1.0.post1
fastapi==0.108.0; python_version >= '3.8'
idna==3.6; python_version >= '3.5'
motor==3.3.2; python_version >= '3.7'
passlib==1.7.4
pyasn1==0.5.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pydantic[email]==2.5.3; python_version >= '3.7'
pydantic-core==2.14.6; python_version >= '3.7'
pydantic-settings==2.1.0; python_version >= '3.8'
pymongo==4.6.1; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
python-jose[cryptography]==3.3.0
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0; python_version >= '3.7'
starlette==0.32.0.post1; python_version >= '3.8'
typing-extensions==4.9.0; python_version >= '3.8'
    uvicorn